Fixed a problem with the input streams created by generate_resource
The problem was that since the code created used a computed variable,
it always returned the same stream. This meant that it could only be
called once (since the Swift InputStream does not reset to the start
after closing and re-opening). This fix changes it to be a function
that returns a new stream each time. Since the stream is just a thin
wrapper over a data buffer, this is still very efficient.